Add revealed votes in `votesPerCandidate.myVotes` Even those which the attempts is no longer in the local storage
