+++ packages/ui/src/council/hooks/useCommitment.ts
@@ -12,11 +12,6 @@ export interface VotingAttempt {
   optionId: string
 }
 
-export interface MyCastVote extends VotingAttempt {
-  voteId?: string
-  isRevealed: boolean
-}
-
 export const useCommitment = (accountId: string, candidateId: string) => {
   const { candidate } = useCandidate(candidateId)
