+++ packages/ui/src/council/hooks/useStoredCastVotes.ts
@@ -1,32 +1,13 @@
 import { useEffect, useMemo } from 'react'
 
-import { useMyAccounts } from '@/accounts/hooks/useMyAccounts'
-import { useLocalStorage } from '@/common/hooks/useLocalStorage'
 import { isDefined } from '@/common/utils'
-import { calculateCommitment } from '@/council/model/calculateCommitment'
 import { useGetCouncilVotesCommitmentsLazyQuery } from '@/council/queries'
 
-import { VotingAttempt } from './useCommitment'
+import { useMyVotingAttempts } from './useMyVotingAttempts'
 
 export const useStoredCastVotes = (cycleId?: number, optionId?: string) => {
   const [get, { data }] = useGetCouncilVotesCommitmentsLazyQuery()
-
-  const [votingAttempts = []] = useLocalStorage<VotingAttempt[]>(`votes:${cycleId}`)
-  const { allAccounts } = useMyAccounts()
-  const myAttempts = useMemo(() => {
-    if (!isDefined(cycleId)) return
-    if (!allAccounts.length) return []
-
-    const addresses = allAccounts.map((account) => account.address)
-    return votingAttempts
-      .filter((attempt) => addresses.includes(attempt.accountId) && (!optionId || optionId === attempt.optionId))
-      .map(({ salt, accountId, optionId }) => ({
-        salt,
-        accountId,
-        optionId,
-        commitment: calculateCommitment(accountId, optionId, salt, cycleId),
-      }))
-  }, [votingAttempts.length, allAccounts.length, cycleId, optionId])
+  const myAttempts = useMyVotingAttempts(cycleId, optionId)
 
   useEffect(() => {
     if (!myAttempts?.length) return
